+++ fs/sdcardfs/packagelist.c
@@ -69,7 +69,7 @@ static unsigned int str_hash(void *key) {
 static int contain_appid_key(struct packagelist_data *pkgl_dat, void *appid) {
         struct hashtable_entry *hash_cur;
 
-        hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist, (unsigned int)appid)
+        hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist, (uintptr_t)appid)
                 if (appid == hash_cur->key)
                         return 1;
 	return 0;
@@ -88,7 +88,7 @@ int get_caller_has_rw_locked(void *pkgl_id, derive_t derive) {
 
 	appid = multiuser_get_app_id(current_fsuid());
 	mutex_lock(&pkgl_dat->hashtable_lock);
-	ret = contain_appid_key(pkgl_dat, (void *)appid);
+	ret = contain_appid_key(pkgl_dat, (void *)(uintptr_t)appid);
 	mutex_unlock(&pkgl_dat->hashtable_lock);
 	//printk(KERN_INFO "sdcardfs: %s: appid=%d, ret=%d\n", __func__, (int)appid, ret);
 	return ret;
@@ -202,7 +202,7 @@ static int insert_int_to_null(struct packagelist_data *pkgl_dat, void *key, int 
 	struct hashtable_entry *new_entry;
 
 	//printk(KERN_INFO "sdcardfs: %s: %d: %d\n", __func__, (int)key, value);
-	hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist, (unsigned int)key) {
+	hash_for_each_possible(pkgl_dat->appid_with_rw,	hash_cur, hlist, (uintptr_t)key) {
 		if (key == hash_cur->key) {
 			hash_cur->value = value;
 			return 0;
@@ -214,7 +214,7 @@ static int insert_int_to_null(struct packagelist_data *pkgl_dat, void *key, int 
 	new_entry->key = key;
 	new_entry->value = value;
 	hash_add(pkgl_dat->appid_with_rw, &new_entry->hlist,
-			(unsigned int)new_entry->key);
+			(uintptr_t)new_entry->key);
 	return 0;
 }
 
@@ -289,7 +289,7 @@ static int read_package_list(struct packagelist_data *pkgl_dat) {
 			while (token != NULL) {
 				if (!kstrtoul(token, 10, &ret_gid) &&
 						(ret_gid == pkgl_dat->write_gid)) {
-					ret = insert_int_to_null(pkgl_dat, (void *)appid, 1);
+					ret = insert_int_to_null(pkgl_dat, (void *)(uintptr_t)appid, 1);
 					if (ret) {
 						sys_close(fd);
 						mutex_unlock(&pkgl_dat->hashtable_lock);
